move ingest client resource around
